+++ cpp/include/copc-lib/io/writer.hpp
@@ -2,6 +2,7 @@
 #define COPCLIB_IO_WRITER_H_
 
 #include <array>
+#include <optional>
 #include <ostream>
 #include <stdexcept>
 #include <string>
@@ -24,9 +25,12 @@ class WriterInternal;
 class Writer : public BaseIO
 {
   public:
-    Writer(std::ostream &out_stream, CopcConfigWriter const &copc_file_writer)
+    Writer(std::ostream &out_stream, const CopcConfigWriter &copc_config_writer,
+           const std::optional<Vector3> &scale = {}, const std::optional<Vector3> &offset = {},
+           const std::optional<std::string> &wkt = {}, const std::optional<las::EbVlr> &extra_bytes_vlr = {},
+           const std::optional<bool> &has_extended_stats = {})
     {
-        InitWriter(out_stream, copc_file_writer);
+        InitWriter(out_stream, copc_config_writer, scale, offset, wkt, extra_bytes_vlr, has_extended_stats);
     }
 
     // Writes the file out
@@ -62,7 +66,10 @@ class Writer : public BaseIO
     };
 
     // Constructor helper function, initializes the file and hierarchy
-    void InitWriter(std::ostream &out_stream, const CopcConfigWriter &copc_file_writer);
+    void InitWriter(std::ostream &out_stream, const CopcConfigWriter &copc_file_writer,
+                    const std::optional<Vector3> &scale, const std::optional<Vector3> &offset,
+                    const std::optional<std::string> &wkt, const std::optional<las::EbVlr> &extra_bytes_vlr,
+                    const std::optional<bool> &has_extended_stats);
     // Gets the sum of the byte size the extra bytes will take up, for calculating point_record_len
     static int NumBytesFromExtraBytes(const std::vector<las::EbVlr::ebfield> &items);
 };
@@ -70,13 +77,16 @@ class Writer : public BaseIO
 class FileWriter : public Writer
 {
   public:
-    FileWriter(const std::string &file_path, const CopcConfigWriter &copc_file_writer)
+    FileWriter(const std::string &file_path, const CopcConfigWriter &copc_file_writer,
+               const std::optional<Vector3> &scale = {}, const std::optional<Vector3> &offset = {},
+               const std::optional<std::string> &wkt = {}, const std::optional<las::EbVlr> &extra_bytes_vlr = {},
+               const std::optional<bool> &has_extended_stats = {})
     {
 
         f_stream_.open(file_path.c_str(), std::ios::out | std::ios::binary);
         if (!f_stream_.good())
             throw std::runtime_error("FileWriter: Error while opening file path.");
-        InitWriter(f_stream_, copc_file_writer);
+        InitWriter(f_stream_, copc_file_writer, scale, offset, wkt, extra_bytes_vlr, has_extended_stats);
     }
 
     void Close() override;