+++ cpp/include/copc-lib/las/header.hpp
@@ -24,7 +24,8 @@ class LasHeader
     LasHeader() = default;
     uint16_t EbByteSize() const;
     LasHeader(int8_t point_format_id, uint16_t point_record_length, const Vector3 &scale, const Vector3 &offset)
-        : point_format_id_(point_format_id), point_record_length_{point_record_length}, scale(scale), offset(offset){};
+        : point_format_id_(point_format_id), point_record_length_{point_record_length}, scale_(scale),
+          offset_(offset){};
 
     // Constructor for python pickling
     // TODO: Add a CMAKE flag to only compute python-specific code when python is compiled
@@ -32,7 +33,7 @@ class LasHeader
               uint32_t vlr_count, const Vector3 &scale, const Vector3 &offset, uint64_t evlr_offset,
               uint32_t evlr_count)
         : point_format_id_(point_format_id), point_record_length_(point_record_length), point_offset_(point_offset),
-          point_count_(point_count), vlr_count_(vlr_count), scale(scale), offset(offset), evlr_offset_(evlr_offset),
+          point_count_(point_count), vlr_count_(vlr_count), scale_(scale), offset_(offset), evlr_offset_(evlr_offset),
           evlr_count_(evlr_count){};
 
     static LasHeader FromLazPerf(const lazperf::header14 &header);
@@ -45,8 +46,8 @@ class LasHeader
 
     uint8_t PointFormatId() const { return point_format_id_; }
     uint16_t PointRecordLength() const { return point_record_length_; }
-    Vector3 Scale() const { return scale; }
-    Vector3 Offset() const { return offset; }
+    Vector3 Scale() const { return scale_; }
+    Vector3 Offset() const { return offset_; }
     uint64_t PointCount() const { return point_count_; }
     uint32_t PointOffset() const { return point_offset_; }
     uint32_t VlrCount() const { return vlr_count_; }
@@ -84,14 +85,14 @@ class LasHeader
     Box Bounds() const;
 
     // Apply Las scale factors to Vector3 or double
-    Vector3 ApplyScale(const Vector3 &unscaled_value) const { return unscaled_value * scale + offset; }
-    Vector3 ApplyInverseScale(const Vector3 &scaled_value) const { return scaled_value / scale - offset; }
-    double ApplyScaleX(double unscaled_value) const { return unscaled_value * scale.x + offset.x; }
-    double ApplyScaleY(double unscaled_value) const { return unscaled_value * scale.y + offset.y; }
-    double ApplyScaleZ(double unscaled_value) const { return unscaled_value * scale.z + offset.z; }
-    double ApplyInverseScaleX(double scaled_value) const { return (scaled_value - offset.x) / scale.x; }
-    double ApplyInverseScaleY(double scaled_value) const { return (scaled_value - offset.y) / scale.y; }
-    double ApplyInverseScaleZ(double scaled_value) const { return (scaled_value - offset.z) / scale.z; }
+    Vector3 ApplyScale(const Vector3 &unscaled_value) const { return unscaled_value * scale_ + offset_; }
+    Vector3 ApplyInverseScale(const Vector3 &scaled_value) const { return scaled_value / scale_ - offset_; }
+    double ApplyScaleX(double unscaled_value) const { return unscaled_value * scale_.x + offset_.x; }
+    double ApplyScaleY(double unscaled_value) const { return unscaled_value * scale_.y + offset_.y; }
+    double ApplyScaleZ(double unscaled_value) const { return unscaled_value * scale_.z + offset_.z; }
+    double ApplyInverseScaleX(double scaled_value) const { return (scaled_value - offset_.x) / scale_.x; }
+    double ApplyInverseScaleY(double scaled_value) const { return (scaled_value - offset_.y) / scale_.y; }
+    double ApplyInverseScaleZ(double scaled_value) const { return (scaled_value - offset_.z) / scale_.z; }
 
     uint16_t file_source_id{};
     uint16_t global_encoding{};
@@ -103,14 +104,14 @@ class LasHeader
     Vector3 max{};
     Vector3 min{};
 
-    // xyz scale/offset
-    Vector3 scale{Vector3::DefaultScale()};
-    Vector3 offset{Vector3::DefaultOffset()};
-
     // # of points per return 0-14
     std::array<uint64_t, 15> points_by_return{};
 
   protected:
+    // xyz scale/offset
+    Vector3 scale_{Vector3::DefaultScale()};
+    Vector3 offset_{Vector3::DefaultOffset()};
+
     int8_t point_format_id_{6};
     uint16_t point_record_length_{};